// csm.web.app/src/app/shared/misc/enums/key.enum.ts
/*
 * *************************************************************************
 *
 * Copyright: Robert Bosch Power Tools GmbH, 2020
 *
 * *************************************************************************
 */

export enum KeyEnum {
    Alt = 'Alt',
    AltGr = 'AltGraph',
    ArrowDown = 'ArrowDown',
    ArrowLeft = 'ArrowLeft',
    ArrowRight = 'ArrowRight',
    ArrowUp = 'ArrowUp',
    Backspace = 'Backspace',
    Control = 'Control',
    Delete = 'Delete',
    Enter = 'Enter',
    Escape = 'Escape',
    Meta = 'Meta',
    Shift = 'Shift',
    Tab = 'Tab',
}